-- Closes ticket #192.
Adds pipeline as a parameter to the user-specified module signature
generator, so that it can take connections into account.

Adds user-specified caching to VTK package. 

This slightly changes the interface to the user-defined caching API,
so if we want to backport this to the v1.0 branch, we are going to
have to live with some minor form of backwards incompatibility.
--

core/cache/hasher.py:

	Hasher.compound_signature: new method

core/interpreter/cached.py:

	Fixed minor comment typo

core/modules/module_registry.py:

	uncommented out parts related to user-defined hashing.

	ModuleRegistry.module_signature: now takes a module and a
	pipeline as parameters, to pass the pipeline to user-defined
	hashers.

	ModuleRegistry.add_module: replaced non-working
	cacheCallable parameter with working signatureCallable
	parameter.

core/vistrail/pipeline.py:

	Pipeline.module_signature: passes self to
	registry.module_signature

packages/vtk/__init__.py:

	bumps minor version to 0.9.1

	createModule: sets user-defined hasher that takes connections
	into account.