DataTransformation_webservice.vt:  Change in demo, instead of using File module, the demo now uses HTTPFile module to be able to get the xsl file from the web (the file was uploaded to VisTrails web site).
noaa_webservices.vt:  Demo was extended and also includes a workaround to use the noaa web services to avoid the bug in the PythonSource (when the user tries to use a port that have different namespace but equal type).  It receives the response object in the PythonSource and gets the values of the objects contained in this.

__init__.py:  When the stub files could not be generated because of a syntax error in the WSDL, the message was only shown in the console, it is now shown to the user in the GUI when the web service package is initialized.