Two small bugfixes for command-line switches:
(1) The -n/--nosplash switch is now working correctly.

(2) The help text for the -m/--movies switch was incorrectly formatted.  This has been fixed.

Note: Carlos and Huy inform me that bug (1) above might be related to the same timing bug that is currently affecting the -x and -s switches.  So we'll want to do some regression testing after the -x/-s bug is fixed to make sure -n still works.