Changed mac distribution files so they work when building the application on Leopard. My old script was Tiger only and I don't have a intel mac with tiger anymore. This means our future distributions for intel will work on Leopard only. Notice that as long as we don't have a library upgrade, I can always update the old bundle with the new vistrails file.
I am also checking in helper scripts for including itk on the distribution. Just to clarify that I don't include itk in the standard distribution, but as I built a version to a user including it in the past, I will keep the files in the repository in case other users request it.

This commit message updates both trunk and branch