I forgot to include these files when committed in revision 1298. I am repeating relevant parts of the commit message just for convenience.
Bug fixes and improvements to the VisTrails extensions for SciVis class:
- Added support to point to a tag instead of a version. This will cause 
  vistrails to be executed every time the wiki page is edited or the latex 
  file is compiled so the document has always the latest image.
- Updated the latex README file to include the new features

extensions/http/run_vistrails.php
 Checking if the tag attribute is present and forcing vistrails to be always executed if it is.

extensions/mediawiki/vistrailsExtension.php
Checking if the tag attribute is present and forcing vistrails to be 
always executed if it is. 
Also checking for new execute and showspreadsheetonly attributes and passing 
to the url for downloading the file so the .vtl file is built accordingly.

extensions/mediawiki/download.php
Checking if the tag, execute and showspreadsheetonly keys are passed in the url and building the .vtl file accordingly

extensions/latex/vistrails.sty
Renamed \vistrails command to \vistrail
Added support to new attributes: tag, showspreadsheetonly and execute
Fixed bug when no options to includegraphics were passed.

extensions/latex/example.tex
Using the new tags and the \vistrail command.

extensions/latex/includevistrail.py
Checking if the tag attribute is present and forcing vistrails to be 
always executed if it is.
Also checking for new execute and showspreadsheetonly attributes and passing 
to the url for downloading the file so the .vtl file is built
accordingly.

vistrails/gui/version_prop.py
       ﻿QVersionEmbed.__init__()
       Renamed  \vistrails command to \vistrail