A few more changes in the control flow package. Two examples were added inside the examples fold, to show the functionality of the Map and the Filter operators. The example with the If operator is already being done.
- packages/controlflow.py
Instead of using "if connector.obj.inputPorts.has_key(name)", we use 
"if name in connector.obj.inputPorts"; instead of using "range", we use 
"xrange"; the control modules now inherits the NotCacheable module too.
- examples/triangle_are.vt
In this example, we calculate a serious of isosurface areas (using the 
Map operator), and then we create a 2D plot with all the areas.
- examples/DDBJ_webService.vt
The aim of this workflow is to, given a human gene, find the species 
that have similar genes. Initially, we use the Map operator; then, in 
another version of the workflow, we use the SimilarityFilter, in order 
to show, in the spreadsheet, only the species that have a similarity 
'score' above 98.00.