A bug was fixed inside the controlflow package. More specifically, if more than one module had connection with an output port of Map, this control structure would be executed more than once; it's fixed now. The If module was changed too. It was returning, as a result, the module chosen in the condition. Now, it returns the results of the output ports of this module that the user wants. A function to register a control module was defined too, so the fringe can be automatically changed through it. Besides, a new module was added, AreaFilter, which is used now inside triangle_area.vt.
The example triangle_area.vt has now a new version, using the AreaFilter 
module.

- packages/controlflow.py
class Fold, function updateFunctionPort(): an "if not self.upToDate" was 
added to fix the bug;
class AreaFilter: a new class added;
class If: new input ports defined to return the results of the output 
ports, as previously said;
function registerControl(): a new function added.

- examples/triangle_area.vt
A new version added to this example.