<ticket>#270</ticket> <ticket>#272</ticket> <bugfix>More robust handling of errors when loading subworkflows</bugfix> <bugfix>Check to ensure that abstraction files actually exist when opening and saving vistrails</bugfix>
This revision fixes some of the subworkflow (abstraction) bugs that
we've run into recently.  Specifically, it works to handle errors when
the abstraction package version changes as well as checking to ensure
that abstraction files exist when opening and saving vistrails.  If an
abstraction exists in memory but the vistrail containing the file is
closed, other vistrails that use that abstraction (through copy-paste,
or two files with the same abstraction) must serialize the abstraction
for the save to work correctly.  In addition, the paste operation
needs to maintain correct descriptors for pasted modules.

core/modules/module_registry.py:
  - ModuleRegistry.delete_descriptor: if package is None and the
    descriptor has package information, use it.

core/vistrail/controller.py:
  - VistrailController.find_abstractions: if abstractions are missing
    the correct package version, use get_similar_descriptor to obtain
    a descriptor that should work.
  - VistrailController.check_abstraction: new method as part of a
    refactoring of ensure_abstractions_loaded, now recursive
  - VistrailController.ensure_abstractions_loaded: refactor out
    check_abstraction for better reuse, recursion
  - VistrailController.write_vistrail: serialize abstractions if they
    exist on memory but the location they were loaded from does not
    exist any longer

db/services/io.py:
  - save_vistrail_to_zip_xml: don't prepend 'abstraction_' when an
    abstraction file already has that prefix, esacalate Exceptions to
    critical status

gui/vistrail_controller.py:
  - VistrailController.paste_modules_and_connections: call
    ensure_modules_are_on_registry to make sure that the correct
    modules descriptors are set when we have package version
    mismatches