<feature>Enable more intuitive command-line interactions with pipelines</feature>
Some preliminary support for command-line interaction with modules,
connections, and functions.  Specifically, allows you to add modules
via a python constructor, set connections and functions via python
method-style syntax, and obtain references to selected items.
Currently, supports setting connections and functions via kwargs to
the constructor, method call syntax, and setattr syntax.  There are
two calls that are added to the locals() of the shell:

  - load_package(identifier: str) -> object: returns a wrapper object that
    allows you to "construct" VisTrails modules from the command-line.

  - selected_modules() -> list(wrapper_module): returns wrapper
    objects that represent VisTrails modules in the command-line

For example, a shell session might look like following:

>>> vtk = load_package('edu.utah.sci.vistrails.vtk')
>>> vtk.vtkDataSetReader()
>>> a = selected_modules()[0]
>>> b = vtk.vtkContourFilter()
>>> b.SetInputConnection0(a.GetOutputPort0())
>>> basic = load_package('edu.utah.sci.vistrails.basic')
>>> c = basic.Integer(value=3)
>>> b.GenerateValues(c.value(), 12.34, 90.98)
>>> a.SetVectorsName = 'my_vectors'
>>> c.value = 5
>>> d_list = [basic.String(value=("test%d" % i)) for i in xrange(3)]
>>> kwargs = dict(("str%d" % (i+1), s.value()) for i, s in enumerate(d_list))
>>> e = basic.ConcatenateString(**kwargs)

The nice thing is that you can use python to automate repetitive
connections, etc (like the last three lines).  Also, note that methods
like the "GenerateValues" call above will do Tuple generation
automatically so the output port "c.value" is connected to the first
input port of a Tuple module of type (Integer, Float, Float) and the
other arguments are set as functions on the remaining input ports.
The output port of the Tuple Module is connected to GenerateValues.

The implementation is accomplished by hacking python's __getattr__,
__new__, __init__, and __setattr__ calls to interact with the
VisTrails api.  It may be possible to improve this handling in the
future.  Currently, everything is dynamically looked up on the
registry so tab-completion (if it were available on the shell via
rlcompleter) wouldn't work, but we could do so in the future.  In
addition, a new object (and type) is created for every call which
probably isn't the right way of doing things.

TODO: Note that all modules are created at (0,0) which is rather
annoying.  Eventually, it would be nice to place modules in available
"open space", and if they connect to other modules already on the
canvas, intuitively close to those modules.  Also, it would be good to
have nicer syntax that looks more like python import for packages
instead of the load_package(<identifier>) or load the wrappers
automatically when the package is enabled.

api/__init__.py:
  - add_module: return the copy of the module that is actually on the
    pipeline, not the one that was initially created
  - add_module_from_descriptor: new method
  - add_connection: return the copy of the connection that is actually
    on the pipeline, not the one that was initially created
  - get_seleted_modules: new method
  - add_port_spec: new method

gui/shell.py:
  - vistrails_port: new class to wrap core.vistrail.port_spec
  - vistrails_module: new class to wrap core.vistrail.module
  - QShell.load_package: new method that is available in the shell
  - QShell.selected_modules: new method that is available in the shell
  - QShell.reset: add load_package and selected_modules to locals

gui/vistrail_controller.py:
  - get_selected_item_ids: new method