<ticket>#286</ticket> <bugfix>Fixed diff so that pipelines created with different versions of packages can try to work correctly.</bugfix>
This bug was caused by the fact that one or more pipelines used older
versions of the ALPS package.  The diff did not attempt to update
package versions if a compatible version could be found in the
registry.  This was fixed by calling ensure_modules_are_on_registry on
these pipelines.

gui/vis_diff.py:
  - QVisualDiff.createDiffPipeline: call
    enusre_modules_are_on_registry on p1 and p2 to update modules that
    may have used older package versions.