<ticket>#301</ticket> <bugfix>Moving condition check to the right place so FileOpen events are captured correctly on a Mac</bugfix>
Running vistrails non-interactively without providing an input file raises only a warning and not an error. This allows for starting a application without the Builder and keep it running. This makes easier for users that want to use vistrails as a library and don't want to implement their own QApplication.

gui/application.py
 - Removed unnecessary imports
 
 - noninteractiveMode()
   Changed error message to warning and returning True in case no input is provided.

 - createWindows()
   installing the event filter if the system platform is a mac

 - eventFilter()
   if the useMacBrishedMetalStyle option is on, then change widgets' attributes.