<feature>Initial version of a persistence package that aims to cache files across VisTrails sessions.</feature>
!!! This feature has had little to no testing and is not complete. !!!

Adds an initial version of persistence that includes a PersistentFile
module along with the necessary support in core and gui.  The idea is
that users can pass any File though the PersistentFile module, and
VisTrails will cache that file (similar to HTTPFile) in the user's
.vistrails directory.  This is accomplished by using the signature of
the upstream subpipeline to identify each file.  Thus, if a user
changes a parameter in a module upstream, PersistentFile will
recompute the upstream subpipeline, but if nothing has changed, it
will use the cached version.  The cache persists across VisTrails
sessions, and supercedes the in-memory cache.  Users can identify
modules that were not executed because of persistence by a blue color.

Eventually, we will need to have methods to reexecute a pipeline even
if a PersistentFile exists.  We will also need methods for keeping the
size of the cache under control.

gui/pipeline_view.py:
  - QPipelineScene.event: add persistence status and use in syntax
  - QPipelineScene.set_module_persistent: new method

gui/theme.py:
  - DefaultTheme.__init__: add PERSISTENT_MODULE_BRUSH

core/interpreter/cached.py:
  - CachedInterpreter.setup_pipeline: set signature on summoned module
  - CachedInterpreter.execute_pipeline: move dictionaries up, store
    cached information
  - CachedInterpreter.finalize_pipeline: set modules to persistent if
    cached
  - CachedInterpreter.add_to_persistent_pipeline: set signature on
    modules

core/modules/vistrails_module.py:
  - Module.__init__: set signature to None

vistrails/packages/persistence:
  - new package to support persistence

vistrails/packages/persistence/__init__.py:
  - PersistentFile: new module
  - initialize: setup directories and configuration