Created simplified drawing functions for workflows and vistrails so that while the user zooms and pans, the graphs get drawn as white boxes without links.
This dramatically improves performance and responsiveness, mostly for the sake of Firefox users.

This commit also contains some refactoring and minor bug fixes.  Most notably, the action names written on the nodes of VT trees used to word wrap if the node crossed the edge of the screen.  Now they get printed offscreen in the same way as the stand alone application.