<ticket>#326</ticket> <bugfix>Modules from different packages or namespaces no longer match</bugfix> <bugfix>Groups try to match on their tag during heuristic matches</bugfix> <feature>Visual Diff now shows matches in white; modules shared by history are colored gray while modules shared by matching are colored white.</feature> <feature>Added better default descriptions to version tree.</feature>
Bunch of fixes related to the visual diff.  Specifically, we resolve
the issue where the heuristic matching in the diff was not taking
namespace or package identifiers into account.  Second, we added some
support for matching groups in the diff.  We are still **not**
matching their underlying pipelines as we should be, but we trying to
match their tags.  Finally, we display matches in the diff separate
from "shared" modules.  Shared modules are now those were created in
an action that is a common ancestor of both selected versions.
Matched modules are those created in actions below the common ancestor
but are similar according to a heuristic match.  Also, some updates to
the default descriptions in the version tree.

core/db/io.py:
  - get_workflow_diff: update for matched pairs
  - get_workflow_diff_with_connections: update for matched pairs and
    fixed a typo

core/vistrail/group.py:
  - Group.add_port_spec: override Module.add_port_spec to use local
    stores instead of db storage since Group doesn't store port specs
  - Group.delete_port_spec: override Module.delete_port_spec to use
    local stores instead of db storage since Group doesn't store port
    specs

core/vistrail/module.py:
  - Module.__str__: include package and namespace information
  - Module.__eq__: check package and namespace information

core/vistrail/vistrail.py:
  - Vistrail.getPipelineDiffByAction: **removed**, obselete
  - Vistrail.get_pipeline_diff_with_connections: call core.db.io
  - Vistrail.getPipelineDiff: renamed to get_pipeline_diff
  - Vistrail.get_pipeline_diff: renamed from getPipelineDiff, removed
    extra obselete code after return
  - Vistrail.get_description: updated to check for annotations, ports
    and more plural descriptions
  - TestVistrail.test2: getPipelineDiff -> get_pipeline_diff

db/services/vistrail.py:
  - heuristicModuleMatch: check namespace and package as well as group
    descriptions
  - getWorkflowDiff: keep matches separate, don't differentiate
    between matches with and without parameter changes, return updated
    tuple with matches

gui/theme.py:
  - DefaultTheme.__init__: add VISUAL_DIFF_MATCH and
    VISUAL_DIFF_SUMMARY brushes

gui/vis_diff.py:
  - QLegendBox: add matched legend entry
  - QVisualDiff.__init__: getPipelineDiff -> get_pipeline_diff
  - QVisualDiff.moduleSelected: getPipelineDiff -> get_pipeline_diff

  - QVisualDiff.createDiffPipeline: updated for matches to be
    displayed separately, fixed bug in adding port specs