<ticket>#259</ticket> <bugfix>Undo and redo are now available when the should be.</bugfix> <bugfix>Execute Version Difference menu item is enabled at the correct times and functions correctly.</bugfix>
Reworked item selection logic so that everything is in the
QGraphicsVersionItem.itemChange since
QVersionTreeScence.mouseReleaseEvent did not covert the non-drag
multiple selection case correctly (release is called before itemChange
in this case).  Note that this may be less efficient; the deselect of
20 versions involves 20 itemChange calls, however, the logic should be
easier to handle and always work.  There are optimizations at the
controller level that ensure that scenes are not regenerated all the
time.

Also updated the versionSelected logic so that it correctly handles
events that are the result of pipeline changes.  Specifically, we do
not reset the redo stack on a versionSelected call unless the user
manually clicks a different version.

These changes may introduce side effects.  Please report any strange
behaviors you notice when interacting with the version tree.

gui/builder_window.py:
  - QBuilderWindow.showDiff: allow diffs between the root and any
    version now

gui/pipeline_tab.py:
  - QPipelineTab.updatePipeline: clear the scene if the current
    pipeline is set to None

gui/version_view.py:
  - QGraphicsLinkItem.itemChange: always return False when trying to
    select a link; note we need to keep the selectable flag so that
    mouse press and release events are handled
  - QGraphicsVersionItem.itemChange: rework logic so that it is more
    obvious what is happening; always emit versionSelected and
    streamline twoVersionsSelected logic
  - QGraphicsVersionItem.mousePressEvent: removed versionSelected
    SIGNAL here because a mouse press may be a deselect
  - QVersionTreeScene.mouseReleaseEvent: **removed**, all logic is now
    handled by itemChange

gui/vistrail_view.py:
  - QVistrailView.versionSelected: rework logic so that only byClick
    changes the selected version, flushes move actions, and empties
    the redo_stack; if versionSelected is called as a consequence of
    an action being updated or a manual change_selected_version call,
    it does not do these actions