<ticket>#276</ticket> <ticket>#287</ticket> <ticket>#291</ticket> <feature>Can delete functions/methods by clicking the check box next to the name</feature> <feature>Can define default values for parameters</feature> <feature>Can define labels for parameters</feature>
Added features to allow more interaction with parameters.  Note that
this is therefore only for input ports!  Specifically, users now have
a more visible way to delete functions (groups of parameters).
Developers can specify default values and labels for parameters.  Note
that a label is currently only displayed in the method dropbox, not in
parameter explorations.  Also, these labels are overridden when a user
specifies an alias for the parameter.  The labels and defaults kwargs
to add_input_port are **string** representations of tuples which are
eval-ed to tuples.  If the method has only a single parameter, the
defaults can be a single value as a string.  Note that a developer may
specify that a parameter not receive a name or label with None.  For
example, if a developer wishes to add a default for the second
parameter of the SetXYZ method, they could do the following:

reg.add_input_port(MyModule, "SetXYZ", [Integer, Integer, Integer], 
                   defaults=str((None, '4', None)))

If they only wish to set defaults for a single parameter:

reg.add_input_port(MyModule, "SetX", Integer, defaults=str(3))

Labels work in a similar manner:

def MyModule(Module):
    ...

    _input_ports = [("SetValues", [Integer, Float], {'labels': str(('Index', 'Value'))}), ...]


core/modules/module_registry.py:
  - ModuleRegistry.auto_add_ports: add support for using a dictionary
    of kwargs after port_name, port_signature in the tuple
  - ModuleRegistry.create_port_spec: add support for labels and defaults
  - ModuleRegistry.add_port: add support for labels and defaults
  - ModuleRegistry.add_input_port: add support for labels and defaults

core/vistrail/controller.py:
  - VistrailController.create_params: make sure value is a string
  - VistrailController.create_function: if no parameter values are
    specified and defaults are, use the defaults

core/vistrail/port_spec.py:
  - PortSpec.__init__: add support for defaults and labels
  - PortSpec.do_copy: add support for defaults and labels
  - PortSpec.convert: add support for defaults and labels
  - PortSpec.labels: new property, parse db_labels string into tuple
  - PortSpec.defaults: new property, parse db_defaults string into tuple

db/specs/all.xml:
  - PortSpec.labels: new field
  - PortSpec.defaults: new field

db/versions/v1_0_1:
  - PortSpec.labels: new field
  - PortSpec.defaults: new field
  - regenerated bindings to allow storage of these entities (only used
    when serializing the registry so I'm not bumping the version
    number of the schema)

gui/method_dropbox.py:
  - QMethodDropBox.updateModule: pass module to addFunction
  - QVerticalWidget.addFunction: accept module, not module id, pass
    port spec through to updateFunction call, connect new deleted
    signal at this level
  - QVerticalWidget.delete_form: new method, move delete code out of
    QMethodInputFrom keyPressEvent
  - QMethodInputForm.__init__: support checkbox and connect clicked to
    a method that emits the delete signal
  - QMethodInputForm.wasClicked: new method, emits deleted signal
  - QMethodInputForm.updateFunction: take port_spec now, use
    port_spec.labels to construct labels
  - QMethodInputForm.keyPressEvent: just emit the deleted signal now,
    moved delete code to QVerticalWidget.delete_form
  - QHoverAliasLabel.__init__: added new default_label parameter
  - QHoverAliasLabel.updateText: if an alias isn't set but a default
    label is specified, use it