<feature>Allow persistence package to search for persistent entities in multiple stores</feature>
Adds a new configuration key-value pair named "search_dbs" that allows
the user to specify other paths to search for persistent files.  Note
that local_db is automatically prepended to search_dbs so it will
always be searched first.  The syntax for the search_dbs string is the
str representation of a python list of strings.  In the configuration
dialog in the preferences dialog, this looks like:

['/path/to/search_db1', '/path/to/search_db2']

and in the startup.xml file itself:

<key name="search_dbs"><str value="['/path/to/search_db1', '/path/to/search_db2']"/></key>

packages/persistence/__init__.py:
  - search_dbs: new global
  - configuration: add search_dbs
  - PersistentPath.updateUpstream: loop through all search_dbs instead
    of just local_db
  - PersistentPath.find: remove unnecessary local_db declaration
  - initialize: get search_dbs from configuration and prepend local_db