<ticket>#290</ticket> <feature>Enable package reloading without the need to restart VisTrails.</feature>
Initial changes to the VisTrails package infrastructure to enable
users and developers to update packages and reload them into VisTrails
without needing to restart VisTrails.  There is still work to be done
here, and we are working on further improvements.  The major change
for developers is that we are decoupling package registration and
package initialization.  Specifically, the __init__.py file of a
package will contain registration information and a second file
(currently must be named "init.py" but this could be configurable from
the __init__.py file) will contain initialize method and all
package-specific python imports.  The other change is that
configuration is now passed to the initialize method so that it can be
registered in __init__.py and referenced in init.py.  Note that this
is backwards-compatible so if initialization is in the __init__.py
file, VisTrails will still load the package normally.  However,
reloading and other new features may not work as advertised!

To accomplish the reloading, we need to track all of the python
modules loaded for a particular package.  We do this by hooking into
python's __builtin__.__import__ method and tracking the added modules.
Note that we only track modules that did not already exist in
sys.modules so lower-level modules are not needlessly reloaded.
However, a VisTrails package that imports modules that "belong" to
other packages can hide dependencies.  Thus, we would like to keep
initialization (and all package-specific python imports) separate from
registration so that all initialization occurs according to the
package dependency graph.  Then, when the package is reloaded, we
disable it and all its reverse dependencies, delete all the python
modules from sys.modules, and then reenable the package and its
reverse dependencies.

Currently, finalize, handle_missing_module are all tied to the
registration file __init__.py, but this will probably change, too.
Also, you need to **erase the cache** if you reload a package after
running a workflow.  This, too, should be changed in the future to
happen automatically.  This commit includes a change of the
ImageMagick package to the new initialization setup.

core/modules/module_registry.py:
  - DuplicateModule: new exception class
  - ModuleRegistry.add_module: check if module already exists, and
    don't allow it to be named twice (raise DuplicateModule)
  - ModuleRegistry.initialize_package: call package's initialize
    indirectly, use init_module instead of module to init _modules and
    _subworkflows
  - ModuleRegistry.remove_package: check that package to be removed
    actually exists and raise an exception otherwise

core/modules/package.py:
  - Package.set_defaults: add _init_module, prefix, and
    py_dependencies fields
  - Package.init_module: new property
  - Package.__hash__: new operator
  - Package.__eq__: new operator
  - Package.__str__: new operator
  - Package._override_import: new internal method
  - Package._reset_import: new internal method
  - Package._import: new internal method to shadow __import__
  - Package.get_py_deps: accessor for py_dependencies
  - Package.remove_py_deps: update py_dependencies
  - Package.load: take prefix and existing paths instead of module so
    that we can reload, set prefix after load, use _import, and import
    from prefix instead of packages. and userpackages. if it is
    specified.
  - Package.initialize: new method, shadow __import__ and run
    initialize from init.py, falling back to run from __init__.py
  - Package.unload: new method, delete dependencies from sys.modules
    so that they will be reloaded when import is called again

core/packagemanager.py:
  - PackageManager.init_registry: use core.modules. prefix instead of
    actual module
  - PackageManager.add_package: packageName -> codepath
  - PackageManager.late_enable_package: use prefix_dictionary, not
    package_dictionary, and catch exceptions on initialization errors
    so that we can remove the package from the registry
  - PackageManager.reload_package: new method, disables package and
    reverse deps, then enables the package and its reverse deps
  - PackageManager.initialize_packages: use prefix_dictionary, track
    existing paths during imports
  - PackageManager.get_all_dependencies: new method
  - PackageManager.all_dependencies: new method
  - PackageManager.all_reverse_dependencies: new method

core/startup.py:
  - VistrailsStartup.load_packages: exclude basic_modules and
    abstraction from being added so that we don't have two copies in
    the package manager!
  - VistrailsStartup.installPackages: change abstraction_dict to use
    prefix

db/services/io.py:
  - import MySQL if it exists so that it isn't seen as a dependency in
    other packages
  - get_db_id_from_name: new method

gui/preferences.py:
  - QPackagesWidget.__init__: add reload button
  - QPackagesWidget.populate_lists: clear list before adding
  - QPackagesWidget.enable_current_package: display error if
    initialization fails
  - QPackagesWidget.reload_current_package: new method to reload
    package
  - QPackagesWidget.set_buttons_to_enabled_package: set reload button
  - QPackagesWidget.set_buttons_to_available_package: set reload
    button
  - QPackagesWidget.set_package_information: display exceptions as a
    dialog
  
packages/ImageMagick/__init__.py:
  - updated package for new initialization routines

packages/ImageMagick/init.py:
  - new file, moved most code from __init__.py, updated initialize to
    accept configuration