<feature>New module, RepoSync, has been added to the HTTP package, which enables data files to be synced with an online repository (crowdLabs)</feature>
Use this module like the File module. On execution of a workflow, the RepoSync module will check if the data file parameter is present on the online repository. If not, it will upload that data. 
This module enables users to have their data files uploaded to an online repository. If someone tries to execute a workflow that doesn't have the local data files, but makes use of the RepoSync module, the synced data will automatically be downloaded from the online repository and used.

packages/HTTP/init.py
 - added RepoSync Module
 - completely refactored HTTPFile, making it cleaner, more robust, and handle large files better