** EXPERIMENTAL PACKAGE **
I'm adding the new version of the persistence package to svn under a
different identifier (edu.utah.sci.vistrails.persistence.exp).  DO NOT
use this package for anything close to critical at this point.
Eventually, this will become a new version of the persistence package
when upgrade support is more stable.  This package requires git
binaries and sqlite3 python support.