<ticket>#364</ticket> <bugfix>Fix issue with annotations that have the same key/value pairs</bugfix>
Add a new configuration option for indexing.  You can now prefix the
index field by a '!' character to indicate that the index is not 1-1
(there may be multiple values for a key).  This is needed for action
annotations because tag values are unique (and thus use the index) but
prune values (and others) are not.

db/bin/auto_gen.py:
  - AutoGen.getAllIndices: detect and add back '!'
  - AutoGen.getIndexName: get rid of '!'
  - AutoGen.getIndexKey: get rid of '!'
  - AutoGen.shouldIgnoreIndexDelete: new method
  - AutoGen.generatePythonClass: add logic for catching KeyError
    exception in delete from index that is flagged to ignore it

db/versions/v1_0_2/specs/all.xml:
  - Vistrail.actionAnnotations: key:value index is now flagged (!)

db/versions/v1_0_2/domain/auto_gen.py:
  - changes from spec/auto_gen changes