<ticket>#367</ticket> <bugfix>Translate old parameter type serialization to match port/portSpecs</bugfix>
Look for parameters that have types <identifier>:<namespace>|<module>
and convert them to <identifier>:<module>:<namespace>.  The issue is
that detecting <identifier>:http://www/address|<namespace>|<module>
requires determining that http is not a module name.  We do this by
checking for "http://" and assuming that http is not the module name.

db/versions/v1_0_2/translate/v1_0_1.py:
  - translateVistrail.update_type: new submethod
  - translateVistrail: update DBParameter.type to match new format