<feature>HTTPFile now supports proxies.</feature> <bugfix>Fixed error when parsing the date from the file's header.</bugfix>
Changed the library used in HTTPFile to give support to proxies. The library
now being used obtain from the O.S. the proxy settings.
For some files the date format in it's header was different. A try/except
statement was added to support both date format.

packages/HTTP/init.py:
 - HTTPFile.compute: changed the library used. From httplib to urllib2
 - HTTPFile.is_cacheable: removed redundant code.
 - HTTPFile._is_outdated: supports different format of date.