Ability to read Config File added to vt server
Added load_config method to application_server.py
This method loads server config data from a file so that it doesn't need
to be hard coded. Passing a config file is required. If it is missing,
incomplete, or unreadable, either a skeleton config file is created or
the missing fields are added for later population.