<ticket>#378</ticket> <bugfix>Fixed logging so that executions are not duplicated when saving more than once per session</bugfix>
Delete the workflow executions on the log after saving them.  Because
we're now appending to the log, we need to delete the executions so
they are not appended twice.  Also fixed an issue where the execution
ids are changed (when saving) but never reset so the id index is
incorrect.

core/log/log.py:
  - Log.delete_all_workflow_execs: new method

core/vistrail/controller.py:
  - VistrailController.write_vistrail: delete log entries from memory
    after they have been persisted to disk

db/services/io.py:
  - save_log_to_xml: save db_id and add it back after persisting (note
    that all ids will be written to disk as -1)