<ticket>#331</ticket> <bugfix>"import userpackages" fixed by reordering init steps</bugfix>
I have not found the cause of the problem but I think we now can
recover from it.  I manually removed the entry for
userPackageDirectory in the config file.  VisTrails then crashes
because it cannot set the path before loading userpackages.

The problem was that setupDefaultFolders(), which creates the
configuration key, was called after the module was loaded. The
solution was to make the check earlier in the startup.  The code is
fairly complex and the order is often important so I hope I did not
break anything.

core/startup.py:
- VistrailsStartup.init(): removed self.setupDefaultFolders()
- VistrailsStartup.__init__(): added self.setupDefaultFolders()