<ticket>#302</ticket> <bugfix>Fixed logic so moved modules are properly detected and positions updated.</bugfix>
Tracked down the annoying move module bug.  The location of the module
was copied to the graphics item but was not updated when the scene
position changed.  We used to compare the persisted location with this
original location instead of the current scenePos, but this has been
changed so we compare against the scenePos now.

gui/pipeline_view.py:
  - QPipelineScene.setupScene: check for moved modules using their
    scenePos instead of the "center" property