Refactored http modified header error handling
When parsing last-modified http headers in the _is_outdated method of
HTTPFile class, an explicit ValueError is caught if
datetime.datetime.strptime fails. Two different string formats were
used, but now if both of those fail, the _is_outdated returns True
instead of raising an error. I made the assumption that if unable to
parse the modified date, we should re-download the remote version.