<feature>Add pinch gesture support to graphics views</feature>
On the Mac, pan gestures (two finger scrolling) are supported by
default on scroll areas, but pinch gestures are not.  This commit uses
Qt's support for gestures to allow zooming using the pinch gesture.
This has only been tested on the Mac so far but should work for other
platforms...

gui/graphics_view.py:
  - QInteractiveGraphicsView.__init__: add gesture inits
  - QInteractiveGraphicsView.viewportEvent: override default method to
    support pinch zooming