<bugfix>With delayed upgrades, moving modules after an upgrade now generates actions in the correct order</bugfix> <ticket>#386</ticket>
When using delayed upgrades, a user can move modules in an upgraded
workflow, but the two actions (upgrade, move) will not be persisted
correctly---only the move is persisted.  That move will be invalid if
the moved module has been upgraded.  This is now fixed by checking to
see if any modules have moved and flushing all delayed actions if so.
This ensures that temporary upgrades are not persisted, but if a
module is moved, the upgrade is persisted before the moves.

core/vistrail/controller.py:
  - VistrailController.has_move_actions: new method

gui/pipeline_view.py:
  - QPipelineScene.hasMoveActions: new method

gui/vistrail_controller.py:
  - VistrailController.has_move_actions: new method

gui/vistrail_view.py:
  - QVistrailView.versionSelected: flush all delayed actions if there
    are move actions