<bugfix>Fixed issue where named intermediate persistent files/directories were not reused correctly</bugfix> <bugfix>Fixed git hash retrieval for persisted directories</bugfix>
When saving a named persistent intermediate output, we were not
correctly determining whether the persistent copy could be used.  This
was due to a number of issues including improper logic for two of the
db calls and the need for a method to retrieve the latest version of a
persistent entity.

vistrails/packages/persistence/db_utils.py:
  - DatabaseAccess.get_signature: typo in sql: verison -> version
  - DatabaseAccess.ref_exists: if version is None, don't put it in the
    where clause, also switch back to boolean return value

vistrails/packages/persistence/init.py:
  - PersistentPath.git_get_hash: updated to allow versioned hash
    access and made this work for directories
  - PersistentPath.git_get_latest_version: new method to retrieve the
    latest version of a file/directory
  - PersistentPath.updateUpstream: if a version is not specified, need
    to get the latest version before checking signature
  - PersistentPath.compute: update to call git_get_hash with path_type