<bugfix>Flush delayed actions (including upgrades) to fix issues with delayed upgrades and grouping upgraded workflows</bugfix>
Note that this keeps some odd behavior when editing version metadata.
The delayed actions will be flushed before tags, descriptions are
updated, it seems like the flush of the version tree isn't happening
quite correctly.  If you need to untag or change descriptions of an
upgraded version, right-click and use the "Show Raw Version" to select
the old version.

gui/vistrail_controller.py:
  - VistrailController.create_group: flush_move -> flush_delayed
  - VistrailController.create_abstraction: flush_move -> flush_delayed
  - VistrailController.create_abstraction_from_group: flush_move ->
    flush_delayed
  - VistrailController.ungroup_set: flush_move -> flush_delayed
  - VistrailController.update_notes: flush_move -> flush_delayed
  - VistrailController.execute_current_workflow: flush_move ->
    flush_delayed
  - VistrailController.update_current_tag: flush_move -> flush_delayed
  - VistrailController.copy_modules_and_connections: flush_move ->
    flush_delayed
  - VistrailController.paste_modules_and_connections: flush_move ->
    flush_delayed