<feature>Add preference for migrating tags and notes on upgrades</feature>
This optional feature will migrate tags and descriptions from an
outdated version to the upgraded version when an upgrade is required.
This is set to False by default so to enable it, go to Expert
Configuration in Preferences and set "migrateTags" to True.

core/configuration.py:
  - default.base_dir: add migrateTags setting

core/vistrail/controller.py:
  - VistrailController.flush_delayed_actions: add logic for migrating tags